thread-safety fixes to debug.cpp - most functions were already reentrant, just CAS-ified the assert suppression flags. (NB: the debug_filter stuff remains non-reentrant because it's usually only called from main() or at least the main thread) closes #654
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@8543 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/res/graphics/ogl_tex.cpp
@@ -1022,7 +1022,7 @@ LibError ogl_tex_bind(Handle ht, size_t unit)
 	// we don't bother catching that and disabling texturing because a
 	// debug warning is raised anyway, and it's quite unlikely.
 	H_DEREF(ht, OglTex, ot);
-	ot->tmu = unit;
+	ot->tmu = (u8)unit;
 
 	// if 0, there's a problem in the OglTex reload/dtor logic.
 	// binding it results in whiteness, which can have many causes;
